requirements.txt
requests 
python-dotenv 
langchain
beautifulsoup4
lxml
tqdm
beautifulsoup4
lxml
tqdm